fix float16 problems due to small native int

+++ float16.h
@@ -52,10 +52,10 @@ extern float from_float16(float16 value);
 #define SIGN(VALUE) (uint8_t)(BITS(VALUE)>>31)
 #define EXPONENT(VALUE) (uint8_t)(BITS(VALUE)>>23)
 #define SIGNIFICAND(VALUE) (uint32_t)(BITS(VALUE)&0x7FFFFF)
-#define IMPLICIT_MSBIT (uint32_t)(1<<23)
+#define IMPLICIT_MSBIT (uint32_t)(1UL<<23)
 
 // Macros to reconstruct a positive IEEE single-precision float
-#define FLOAT_BITS(EXP,SIG) ((EXP<<23)|(SIG))
+#define FLOAT_BITS(EXP,SIG) ((((uint32_t)(EXP))<<23UL)|(SIG))
 #define FROM_BITS(VALUE) (*(float*)(&(VALUE)))
 
 #endif
